strip whitespace from channel name
Prior to this commit, the actual channel name passed onto the client
handlers included the `\r\n` from the raw command. Handlers then had to
deal with inconsistent channel name formatting based on whether the
event was something like a `:received` or a `:mentioned` than if the
event was `:joined` or `:parted`. This commit strips leading and
trailing whitespace from the channel name in order to provide consistent
behavior.