Allow ExIrc.Client.start_link to use process opts
Prior to this commit, registering the name for an ExIrc.Client process,
it was a two-step process of `ExIrc.Client.start_link` and then
`Process.register`. This also made registering a name for
supervisor-started `ExIrc.Client` processes a bit tricky. This commit
allows end users to start `ExIrc.Client` processes from a supervisor,
but still register a name, as in the example below.

```
    children = [
      worker(ExIrc.Client, [[debug: true], [name: :irc_client]])
      ]
    supervise(children, strategy: :one_for_all)
```