Relax Elixir version requirement
Currently, ExIrc specifies its Elixir version to `1.0.0`. This will result in a warning message when compiling on anything newer than `1.0.x`, e.g. the latest elixir master, which is currently `1.1-dev`. Relaxing this version requirement will remove this warning message, and since Elixir follows Semantic Versioning, this should not be an issue.