handle QUIT messages
Prior to this commit ExIrc did not know what to do with QUIT messages,
which would result in nicks remaining in the `Channels` data structure
when they should not. This commit implements handling of the QUIT
message and ensures that the new `Channels.user_quit` function is called
to flush the departing user from all known channels.

Close #40