Add test for failing RPL_TOPIC parse
Context: I receive these messages from the Slack IRC bridge in the event
that a given channel has no topic set.  I would like for ExIrc to avoid
crashing in this case.

As far as I can tell, they are not RFC-compliant messages.  From [RFC
2812](https://tools.ietf.org/html/rfc2812):

```
       331    RPL_NOTOPIC
              "<channel> :No topic is set"
       332    RPL_TOPIC
              "<channel> :<topic>"

         - When sending a TOPIC message to determine the
           channel topic, one of two replies is sent.  If
           the topic is set, RPL_TOPIC is sent back else
           RPL_NOTOPIC.
```

Specifically, these are cases where the Slack channel has no topic set.
However, they are sending "channel :" in a 332.

It shouldn't be possible to end up with this result in an RFC-compliant
server, as attempting to set the topic to the empty string should
instead mark the topic as unset.  Consequently:

1) This is a bug in Slack's IRC bridge implementation.
2) I don't care because I would like ExIrc to be able to connect to
slack without falling over dead.