Add section of projects that use ExIrc
By adding this change to the README, I hope to get a comphrensive list
of projects that use ExIrc. The other thing that this commit helps with
is for newcomers to the ExIrc project. By keeping a list of projects
that use ExIrc, we can help newcomers get a "grasp" of how ExIrc works,
too!